+++ NEWS.md
@@ -45,6 +45,7 @@ polmineR 0.7.11.9001
 * Counting over a `partition_bundle` started to throw a warning that an argument arrives at the `cpos()`-method that is not used. The cause for the warning message is removed, an additional unit test has been introduced to recognize issues with the `count`-method (#90).
 * The `kwic()`-method threw an error when trimming the matches by using a positivelist or a stoplist resulted in no remaining matches. The method will now return a NULL object and keep issuing a warning if no matches remain after filtering (#91).
 * Chaining subsetting calls on a corpus/subcorpus omitted filling the s_attribute slot of the `subcorpus` object, resulting in false results when counting over subcorpora. Fixed.
+* Implemented possibility to use multiple queries in `dispersion`-method (#92).
 
 
 polmineR 0.7.11